refactor(Function): 重构小车运动控制逻辑
- 修改了 TaskRock.py 和 TaskRocket.py 中的小车运动控制逻辑
- 优化了小车移动方向，提高了对障碍物位置的判断准确性
- 新增了 TaskWater.py 文件，实现了水上任务的抓取和移动逻辑
